codegen: Use a docker container for generating keysyms
Previously, keysyms were just generated on the host operating system.
This means that the keysyms that are generated varies from system to
system, which is just bad form. This commit makes it so the keysyms are
generated inside of a Docker container, which should make the process
much more reproducible.

Signed-off-by: John Nunley <dev@notgull.net>